Updated EncoderABC to work with float values

+++ java/odometry-abc/src/test/java/Main.java
@@ -7,7 +7,7 @@ final class EncoderTest extends EncoderABC {
     }
 
     @Override
-    protected int read_value(){
+    protected float read_value(){
         return 1;
     }
 }
